Add new interface and new client with delegate
